+++ src/implementation/Server/GRPCServer/GRPCServerImpl.ts
@@ -394,7 +394,6 @@ export default class GRPCServerImpl implements IAppCallbackServer {
     return callback(null, res);
   }
 
-  // @todo: WIP
   async onTopicEvent(
     call: grpc.ServerUnaryCall<TopicEventRequest, TopicEventResponse>,
     callback: grpc.sendUnaryData<TopicEventResponse>,
@@ -420,10 +419,9 @@ export default class GRPCServerImpl implements IAppCallbackServer {
 
     const data = Buffer.from(req.getData()).toString();
     let dataParsed: any;
-
     try {
       dataParsed = JSON.parse(data);
-    } catch (e) {
+    } catch (_) {
       dataParsed = data;
     }